Write repeat_first_field into re-encoded MPEG-2, and keep the sequence interlaced
libavcodec's MPEG-2 encoder never writes repeat_first_field, so a
re-encoded stretch of film came out two fields a picture: a fifth short in
the stream's own terms, with a field parity that no longer alternated into
the copied pictures after it. On a recording whose first picture is film,
the container reports a progressive field order and the re-encoded stretch
also announced progressive_sequence=1 among pictures saying 0.

A picture may repeat a field only when coded as a whole frame, which
libavcodec does only in a progressive sequence. So a segment with a
repeated field in it (read off the packets before the encoder opens) is
encoded as one, and each picture is put back into the recording's
sequence on the way out: the sequence extension gets the recording's
progressive_sequence, the picture its own field order and repeat. The
macroblocks read the same either way. Broadcasts without repeated fields
are not read for this and cut to the same bytes as before.

tests/run_pulldown_tests.sh checks field count, parity and sequence on
the elementary stream of five cuts; 0.8.6 fails all five.

+++ rust/crates/core/src/weave.rs
@@ -125,7 +125,7 @@ impl Weave {
         // picture's, and come back as that picture again on each frame.
         let fields = 2 + i64::from(unsafe { (*picture.as_ptr()).repeat_pict }.clamp(0, 4));
         let first_top = unsafe {
-            (*picture.as_ptr()).flags & ff::ffi::AV_FRAME_FLAG_TOP_FIELD_FIRST as i32 != 0
+            (*picture.as_ptr()).flags & ff::ffi::AV_FRAME_FLAG_TOP_FIELD_FIRST != 0
         };
         let from = ((t - head) / self.field).round() as i64;
         let picture = Rc::new(picture);
@@ -282,9 +282,9 @@ mod tests {
             let p = f.as_mut_ptr();
             (*p).repeat_pict = fields - 2;
             if top_first {
-                (*p).flags |= ff::ffi::AV_FRAME_FLAG_TOP_FIELD_FIRST as i32;
+                (*p).flags |= ff::ffi::AV_FRAME_FLAG_TOP_FIELD_FIRST;
             } else {
-                (*p).flags &= !(ff::ffi::AV_FRAME_FLAG_TOP_FIELD_FIRST as i32);
+                (*p).flags &= !ff::ffi::AV_FRAME_FLAG_TOP_FIELD_FIRST;
             }
         }
         f
